Refactor project related state functions out of useWorkstationState, fix bug where Project name was not being saved properly
