+++ src/utils/hooks/use-workstation-state.ts
@@ -1,19 +1,33 @@
 import { DiffableState } from "enums/diffable-state";
 import { SetStateAction, useAtom } from "jotai";
-import _ from "lodash";
-import { ProjectRecord } from "models/project-record";
 import { WorkstationStateRecord } from "models/workstation-state-record";
 import { useCallback } from "react";
 import { WorkstationStateAtomFamily } from "utils/atoms/workstation-state-atom-family";
 
 interface UseWorkstationStateResult {
+    /**
+     * **Initial** state value (at time of load, last save, etc.)
+     */
     initialState: WorkstationStateRecord;
+    /**
+     * Returns true if the current state does not equal the initial state
+     */
     isDirty: boolean;
-    setCurrentProject: (update: SetStateAction<ProjectRecord>) => void;
+    /**
+     * Sets the **current** state only
+     */
     setCurrentState: (update: SetStateAction<WorkstationStateRecord>) => void;
-    setInitialProject: (update: SetStateAction<ProjectRecord>) => void;
-    setProject: (updatedProject: SetStateAction<ProjectRecord>) => void;
+    /**
+     * Sets the **initial** state only
+     */
+    setInitialState: (update: SetStateAction<WorkstationStateRecord>) => void;
+    /**
+     * Sets the **initial** and **current** state
+     */
     setState: (update: SetStateAction<WorkstationStateRecord>) => void;
+    /**
+     * **Current** state value
+     */
     state: WorkstationStateRecord;
 }
 
@@ -25,43 +39,6 @@ const useWorkstationState = (): UseWorkstationStateResult => {
         WorkstationStateAtomFamily(DiffableState.Current)
     );
 
-    const _setProject = useCallback(
-        (
-            setWorkstationState: (
-                update: (prev: WorkstationStateRecord) => WorkstationStateRecord
-            ) => void,
-            updatedProject: SetStateAction<ProjectRecord>
-        ) =>
-            setWorkstationState((prev: WorkstationStateRecord) =>
-                prev.merge({
-                    project: _.isFunction(updatedProject)
-                        ? updatedProject(prev.project)
-                        : updatedProject,
-                })
-            ),
-        []
-    );
-
-    const setInitialProject = useCallback(
-        (updatedProject: SetStateAction<ProjectRecord>) =>
-            _setProject(setInitialState, updatedProject),
-        [_setProject, setInitialState]
-    );
-
-    const setCurrentProject = useCallback(
-        (updatedProject: SetStateAction<ProjectRecord>) =>
-            _setProject(setCurrentState, updatedProject),
-        [_setProject, setCurrentState]
-    );
-
-    const setProject = useCallback(
-        (updatedProject: SetStateAction<ProjectRecord>) => {
-            setCurrentProject(updatedProject);
-            setInitialProject(updatedProject);
-        },
-        [setCurrentProject, setInitialProject]
-    );
-
     const setState = useCallback(
         (updatedWorkstationState: SetStateAction<WorkstationStateRecord>) => {
             setInitialState(updatedWorkstationState);
@@ -70,19 +47,13 @@ const useWorkstationState = (): UseWorkstationStateResult => {
         [setCurrentState, setInitialState]
     );
 
-    const isPersisted = state.project.isPersisted();
-    const isProjectDirty = !initialState.project.equals(state.project);
-    const isTrackListDirty = !initialState.tracks.equals(state.tracks);
-
-    const isDirty = isPersisted && (isProjectDirty || isTrackListDirty);
+    const isDirty = !state.equals(initialState);
 
     return {
         initialState,
         isDirty,
-        setCurrentProject,
         setCurrentState,
-        setInitialProject,
-        setProject,
+        setInitialState,
         setState,
         state,
     };
